requirements.txt
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
graphviz==0.15
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.3
neat-python==0.92
networkx==2.5
numpy==1.19.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.9.3
